Check device with blkid before cryptsetup create
Currently in rc.sysinit we check that blkid returns 2 before running
mkswap. This is a little bit stronger. Even without mkswap "cryptsetup
create" can be dangerous, so do the blkid check for all plain (non-LUKS)
mappings. Furthermore, just check for a non-zero return status, if this
is too broad we can get more specific.

Note: A user can still destroy her data if she enters the wrong password
for a cryptsetup create. This we'll leave as the user's fault; she
should be aware of the limitations of a plain mapping and the
--verify-passphrase cryptsetup option.